KorBackend/requirements.txt
Django==2.0.9
djangorestframework==3.9.0
gunicorn==19.7.1
Pillow==3.3.0
Markdown==2.6.9
django-filter==1.0.4